Mission accomplished?

+++ Project/classes/Ijdb/Controllers/Joke.php
@@ -1,20 +1,19 @@
 <?php
 namespace Ijdb\Controllers;
 use \Ninja\DatabaseTable;
+use \Ninja\Authentication;
 
-class Joke
-{
+class Joke {
 	private $authorsTable;
 	private $jokesTable;
 
-	public function __construct(DatabaseTable $jokesTable, DatabaseTable $authorsTable)
-	{
+	public function __construct(DatabaseTable $jokesTable, DatabaseTable $authorsTable, Authentication $authentication) {
 		$this->jokesTable = $jokesTable;
 		$this->authorsTable = $authorsTable;
+		$this->authentication = $authentication;
 	}
 
-	public function list()
-	{
+	public function list() {
 		$result = $this->jokesTable->findAll();
 
 		$jokes = [];
@@ -26,64 +25,88 @@ class Joke
 				'joketext' => $joke['joketext'],
 				'jokedate' => $joke['jokedate'],
 				'name' => $author['name'],
-				'email' => $author['email']
+				'email' => $author['email'],
+				'authorId' => $author['id']
 			];
+
 		}
+
 
 		$title = 'Joke list';
 
 		$totalJokes = $this->jokesTable->total();
 
-		return [
-			'template' => 'jokes.html.php',
-			'title' => $title,
-			'variables' => [
-				'totalJokes' => $totalJokes,
-				'jokes' => $jokes
-			]
-		];
+		$author = $this->authentication->getUser();
+
+		return ['template' => 'jokes.html.php',
+				'title' => $title,
+				'variables' => [
+						'totalJokes' => $totalJokes,
+						'jokes' => $jokes,
+						'userId' => $author['id'] ?? null
+					]
+				];
 	}
 
-	public function home()
-	{
+	public function home() {
 		$title = 'Internet Joke Database';
 
 		return ['template' => 'home.html.php', 'title' => $title];
 	}
 
-	public function delete()
-	{
+	public function delete() {
+
+		$author = $this->authentication->getUser();
+
+		$joke = $this->jokesTable->findById($_POST['id']);
+
+		if ($joke['authorid'] != $author['id']) {
+			return;
+		}
+
+
 		$this->jokesTable->delete($_POST['id']);
 
 		header('location: /joke/list');
 	}
 
-	public function saveEdit()
-	{
-	    $joke = $_POST['joke'];
-	    $joke['jokedate'] = new \DateTime();
-	    $joke['authorId'] = 1;
+	public function saveEdit() {
+		$author = $this->authentication->getUser();
 
-	    $this->jokesTable->save($joke);
 
-	    header('location: /joke/list');
+		if (isset($_GET['id'])) {
+			$joke = $this->jokesTable->findById($_GET['id']);
+
+			if ($joke['authorid'] != $author['id']) {
+				return;
+			}
+		}
+
+		$joke = $_POST['joke'];
+		$joke['jokedate'] = new \DateTime();
+		$joke['authorId'] = $author['id'];
+
+		$this->jokesTable->save($joke);
+
+		header('location: /joke/list');
 	}
 
-	public function edit()
-	{
-	    if (isset($_GET['id'])) {
-	        $joke = $this->jokesTable->findById($_GET['id']);
-	    }
+	public function edit() {
+		$author = $this->authentication->getUser();
 
-	    $title = 'Edit joke';
+		if (isset($_GET['id'])) {
+			$joke = $this->jokesTable->findById($_GET['id']);
+		}
 
-	    return [
-			'template' => 'editjoke.html.php',
-	        'title' => $title,
-	        'variables' => [
-	            'joke' => $joke ?? null
-	        ]
-	    ];
+		$title = 'Edit joke';
+
+		return ['template' => 'editjoke.html.php',
+				'title' => $title,
+				'variables' => [
+						'joke' => $joke ?? null,
+						'userId' => $author['id'] ?? null
+					]
+				];
 	}
 
 }
